Added set_file_watch utility
First argument is the file to watch, can be a list enclosed in "..." but I have only
tested with one file.

Second argument is a command to run.  The script puts that command in a script that will
output a purple line then call the command and exit 0.  This is necessary because the
fswatcher stops if the command it calls returns non-0. This ensures that even if the command
itself fails, the fswatch process sees a 0 and keeps going.