Major overhaul to link-pwd
No more .delete files.  Absence of .restore file is enough to figure out that if
there is a /target/ link, we delete it.

Made link and unlink and restore commands instead of a toggle with a file.

Added possiblity to put an array at the top of the script to use instead of the
linkfile so I can distribute the script to people.