Force adding of in a very an edge case ...
Suppose you do 'cd /a/b/c/<tab>' and a/b/c contains
no directories, and you do backspace so that the
cursor is right after the last '/', then pressing
<tab> again will not add a space because COMPREPLY
will have zero elements.  This forces the space to
happen by putting ${cur} in COMPREPLY but only if
it exists.