Ensure termination of child process in all situations
When a new event happens, the child process needs to be killed but
because there is a level of indirection, (the 'run' function), this
commit needs to be done.

The run function is exported and launched with `setsid bash -c` so that
no signals get propagated to that process.

The run function sets up a trap to kill its child process when it gets a
TERM signal.

When a new event occurs or when the iwatch program exits, a TERM signal
is sent to the run subprocess and it kills its child process.